backend/backend_services/friendship/app/requirements.txt
django
psycopg2
djangorestframework
gunicorn
dotmap
requests